Refactor: 이미지 링크 기반 추천 프롬프트 및 메소드 수정 (#56)
