+++ src/main/java/com/adoonge/seedzip/recommendation/dto/request/ChatGPTRequest.java
@@ -1,9 +1,8 @@
-package com.adoonge.seedzip.summary.dto.request;
+package com.adoonge.seedzip.recommendation.dto.request;
 
-import com.adoonge.seedzip.summary.dto.Message;
-import com.adoonge.seedzip.summary.dto.TextMessage;
+import com.adoonge.seedzip.recommendation.dto.Message;
+import com.adoonge.seedzip.recommendation.dto.TextMessage;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.Collections;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -24,12 +23,12 @@ public class ChatGPTRequest {
     @JsonProperty("max_tokens")
     private int maxTokens;
 
-    public static ChatGPTRequest createImageRequest(String model, int maxTokens, String role, String imageUrl) {
-        String prompt = "Please answer in Korean. Analyze the image from the given URL and extract title, a 3-sentence summary, and tags as a comma-separated string without #, then return the result as a raw JSON object without any code block or additional formatting.";
-        TextContent textContent = new TextContent("text", prompt);
+    public static ChatGPTRequest createImageRequest(String model, int maxTokens, String imageUrl) {
+        Message systemMessage = new TextMessage("system", "You are a helpful assistant for generating hashtags for . Please answer in Korean.");
+        TextContent textContent = new TextContent("text", "Analyze the image from the given URL and extract title, a 3-sentence summary, and tags as a comma-separated string without #, then return the result as a raw JSON object without any code block or additional formatting.");
         ImageContent imageContent = new ImageContent("image_url", new ImageUrl(imageUrl));
-        Message message = new ImageMessage(role, List.of(textContent, imageContent));
-        return createChatGPTRequest(model, maxTokens, Collections.singletonList(message));
+        Message userMessage = new ImageMessage("user", List.of(textContent, imageContent));
+        return createChatGPTRequest(model, maxTokens, List.of(systemMessage, userMessage));
     }
